Renaming create_teams to teamgen
