+++ heltour/tournament/teamgen.py
